moving text2sql metrics functions to sql_utils
