+++ src/unitxt/dataset.py
@@ -14,7 +14,7 @@ from .collections import __file__ as _
 from .collections_operators import __file__ as _
 from .dataclass import __file__ as _
 from .dataset_utils import get_dataset_artifact
-from .db_utils import __file__ as _
+from .sql_utils import __file__ as _
 from .deprecation_utils import __file__ as _
 from .dialog_operators import __file__ as _
 from .dict_utils import __file__ as _
