+++ src/unitxt/sql_utils.py
@@ -6,6 +6,7 @@ from abc import ABC, abstractmethod
 from functools import lru_cache
 from typing import Any, List, Optional
 
+import re
 import requests
 from huggingface_hub import snapshot_download
 from requests.exceptions import ConnectionError, ReadTimeout
@@ -339,3 +340,209 @@ def get_db_connector(db_type: str):
         raise ValueError(f"Unsupported database type: {db_type}")
 
     return connector
+
+
+def is_sqlglot_parsable(sql: str, db_type="sqlite") -> bool:
+    """Returns True if sqlglot does not encounter any error, False otherwise."""
+    from sqlglot import parse
+
+    if db_type == "db2":
+        db_type = "postgres"  ## TODO: temporary until sqlglot adds support for db2
+    try:
+        parse(sql, read=db_type)
+        return True
+    except Exception as e:
+        logger.debug(f"SQL query could not parse: {e}")
+        return False
+
+
+def is_sqlparse_parsable(sql: str) -> bool:
+    """Returns True if sqlparse does not encounter any error, False otherwise."""
+    from sqlparse import parse
+    from sqlparse.tokens import Error
+
+    try:
+        statements = parse(sql)
+        for statement in statements:
+            for token in statement.tokens:
+                if token.ttype == Error:
+                    return False
+        return True
+    except Exception as e:
+        logger.debug(f"SQL query could not parse: {e}")
+        return False
+
+
+def sqlglot_optimized_equivalence(expected: str, generated: str) -> int:
+    """Checks if SQL queries are equivalent using SQLGlot parsing, so we don't run them."""
+    from sqlglot import diff, parse_one
+    from sqlglot.optimizer import optimize
+
+    try:
+        t_diff = diff(
+            optimize(parse_one(expected.lower()).sql(pretty=True)),
+            optimize(parse_one(generated.lower()).sql(pretty=True)),
+        )
+        sql_diff = sum(0 if (e.__class__.__name__ == "Keep") else 1 for e in t_diff)
+
+        return 1 if sql_diff == 0 else 0
+    except Exception as e:
+        logger.debug(f"Error parsing SQL for comparison: {e}")
+        return False
+
+
+def extract_select_info(sql: str):
+    """Parse SQL using sqlparse and return a dict of extracted columns and clauses."""
+    from sqlparse import parse
+    from sqlparse.sql import Identifier, IdentifierList
+    from sqlparse.tokens import DML, Keyword
+
+    statements = parse(sql)
+    if len(statements) != 1:
+        return None
+    stmt = statements[0]
+    if not any(t.ttype is DML and t.value.upper() == "SELECT" for t in stmt.tokens):
+        return None
+    parts = {
+        "columns": None,
+        "from": "",
+        "where": "",
+        "group": "",
+        "having": "",
+        "order": "",
+    }
+
+    def extract_select_columns(statement):
+        columns = []
+        select_seen = False
+        for token in statement.tokens:
+            if token.ttype is DML and token.value.upper() == "SELECT":
+                select_seen = True
+                continue
+            if select_seen:
+                if token.ttype is Keyword and token.value.upper() in (
+                    "FROM",
+                    "WHERE",
+                    "GROUP",
+                    "HAVING",
+                    "ORDER",
+                    "LIMIT",
+                ):
+                    break
+                if isinstance(token, IdentifierList):
+                    for identifier in token.get_identifiers():
+                        columns.append(strip_alias(identifier.value))
+                elif isinstance(token, Identifier):
+                    columns.append(strip_alias(token.value))
+                else:
+                    val = token.value.strip()
+                    if val:
+                        columns.append(strip_alias(val))
+        return frozenset(columns)
+
+    def strip_alias(col: str) -> str:
+        """Remove any AS alias from a column."""
+        col = col.strip()
+        upper = col.upper()
+        if " AS " in upper:
+            return col[: upper.index(" AS ")].strip()
+        parts_alias = col.split()
+        if len(parts_alias) > 1:
+            return " ".join(parts_alias[:-1])
+        return col
+
+    columns = extract_select_columns(stmt)
+    if not columns:
+        columns = frozenset()
+    parts["columns"] = columns
+
+    def collect_clause(statement, clause_keyword):
+        found = False
+        collected = []
+        for token in statement.tokens:
+            tvalue = token.value.upper()
+            if token.ttype is Keyword:
+                if tvalue.startswith(clause_keyword):
+                    found = True
+                    continue
+                if found and tvalue in (
+                    "FROM",
+                    "WHERE",
+                    "GROUP",
+                    "HAVING",
+                    "ORDER",
+                    "LIMIT",
+                ):
+                    break
+            if found:
+                collected.append(token.value)
+        return " ".join(collected).strip()
+
+    parts["from"] = collect_clause(stmt, "FROM")
+    parts["where"] = collect_clause(stmt, "WHERE")
+    parts["group"] = collect_clause(stmt, "GROUP")
+    parts["having"] = collect_clause(stmt, "HAVING")
+    parts["order"] = collect_clause(stmt, "ORDER")
+    return parts
+
+
+def sqlparse_queries_equivalent(sql1: str, sql2: str) -> bool:
+    """Return True if both SQL queries are naively considered equivalent."""
+    try:
+        info1 = SQLNonExecutionAccuracy.extract_select_info(sql1)
+        info2 = SQLNonExecutionAccuracy.extract_select_info(sql2)
+        if not info1 or not info2:
+            return False
+        if info1["columns"] != info2["columns"]:
+            return False
+        for k in ["from", "where", "group", "having", "order"]:
+            if info1[k].replace(" ", "").upper() != info2[k].replace(" ", "").upper():
+                return False
+        return True
+    except Exception as e:
+        logger.debug(f"Errpr parsing SQL query for comparison: {e}")
+        return False
+
+
+def sqlglot_parsed_queries_equivalent(sql1: str, sql2: str, dialect: str = "") -> bool:
+    from sqlglot import exp, parse_one
+
+    try:
+        ast1 = parse_one(sql1, read=dialect)
+        ast2 = parse_one(sql2, read=dialect)
+    except:
+        return False
+    if not (isinstance(ast1, exp.Select) and isinstance(ast2, exp.Select)):
+        return False
+
+    def normalized_select_columns(select_expr: exp.Select):
+        cols = []
+        for item in select_expr.expressions:
+            copy_item = item.copy()
+            copy_item.set("alias", None)
+            cols.append(copy_item.sql(dialect=dialect, normalize=True))
+        return frozenset(cols)
+
+    if normalized_select_columns(ast1) != normalized_select_columns(ast2):
+        return False
+
+    def normalized_clause(expr: exp.Expression, key: str):
+        clause = expr.args.get(key)
+        return clause.sql(dialect=dialect, normalize=True) if clause else ""
+
+    for clause_key in ("from", "where", "group", "having", "order"):
+        if normalized_clause(ast1, clause_key) != normalized_clause(ast2, clause_key):
+            return False
+
+    return True
+
+
+def sql_exact_match(sql1: str, sql2: str) -> bool:
+    """Return True if two SQL strings match after very basic normalization."""
+
+    def normalize_sql(s: str) -> str:
+        s = s.strip().rstrip(";")
+        s = re.sub(r"\s+", " ", s)
+        return s.upper()
+
+    return normalize_sql(sql1) == normalize_sql(sql2)
